It is now allowed to use Settings from other threads

+++ src/proofcore/settingsgroup.cpp
@@ -107,8 +107,11 @@ QSet<QString> SettingsGroup::values() const
 
 SettingsGroup *SettingsGroup::group(const QString &groupName, Settings::NotFoundPolicy notFoundPolicy)
 {
+    SettingsGroup *result;
+    if (ProofObject::call(this, &SettingsGroup::group, Proof::Call::Block, result, groupName, notFoundPolicy))
+        return result;
     Q_D(SettingsGroup);
-    SettingsGroup *result = d->groups.value(groupName, nullptr);
+    result = d->groups.value(groupName, nullptr);
     if (!result && d->globalGroup) {
         result = d->globalGroup->group(groupName, notFoundPolicy == Settings::NotFoundPolicy::AddGlobal
                                                       ? Settings::NotFoundPolicy::Add
@@ -123,6 +126,9 @@ SettingsGroup *SettingsGroup::group(const QString &groupName, Settings::NotFound
 
 QVariant SettingsGroup::value(const QString &key, const QVariant &defaultValue, Settings::NotFoundPolicy notFoundPolicy)
 {
+    QVariant result;
+    if (ProofObject::call(this, &SettingsGroup::value, Proof::Call::Block, result, key, defaultValue, notFoundPolicy))
+        return result;
     Q_D(SettingsGroup);
     if (!d->values.contains(key)) {
         if (d->globalGroup) {
@@ -141,6 +147,9 @@ QVariant SettingsGroup::value(const QString &key, const QVariant &defaultValue, 
 
 SettingsGroup *SettingsGroup::addGroup(const QString &groupName)
 {
+    SettingsGroup *result;
+    if (ProofObject::call(this, &SettingsGroup::addGroup, Proof::Call::Block, result, groupName))
+        return result;
     Q_D(SettingsGroup);
     SettingsGroup *newGroup = d->groups.value(groupName, nullptr);
     if (!newGroup) {
@@ -169,6 +178,8 @@ SettingsGroup *SettingsGroup::addGroup(const QString &groupName)
 
 void SettingsGroup::setValue(const QString &key, const QVariant &value, Settings::Storage storage)
 {
+    if (ProofObject::call(this, &SettingsGroup::setValue, Proof::Call::Block, key, value, storage))
+        return;
     Q_D(SettingsGroup);
     if (storage == Settings::Storage::Global && d->globalGroup) {
         d->globalGroup->setValue(key, value);
@@ -198,6 +209,8 @@ void SettingsGroup::setValue(const QString &key, const QVariant &value, Settings
 
 void SettingsGroup::deleteGroup(const QString &groupName, Settings::Storage storage)
 {
+    if (ProofObject::call(this, &SettingsGroup::deleteGroup, Proof::Call::Block, groupName, storage))
+        return;
     Q_D(SettingsGroup);
     if (storage == Settings::Storage::Global && d->globalGroup) {
         d->globalGroup->deleteGroup(groupName);
@@ -234,6 +247,8 @@ void SettingsGroup::deleteGroup(const QString &groupName, Settings::Storage stor
 
 void SettingsGroup::deleteValue(const QString &key, Settings::Storage storage)
 {
+    if (ProofObject::call(this, &SettingsGroup::deleteValue, Proof::Call::Block, key, storage))
+        return;
     Q_D(SettingsGroup);
     if (storage == Settings::Storage::Global && d->globalGroup) {
         d->globalGroup->deleteValue(key);
@@ -245,6 +260,8 @@ void SettingsGroup::deleteValue(const QString &key, Settings::Storage storage)
 
 void SettingsGroup::copyTo(SettingsGroup *destination)
 {
+    if (ProofObject::call(this, &SettingsGroup::copyTo, Proof::Call::Block, destination))
+        return;
     Q_D(SettingsGroup);
     if (!destination || destination == this)
         return;
